+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidyr (development version)
 
+* Improved the error message returned by `pivot_wider()` when a column selected
+  by `names_from` or `values_from` is also selected by `id_cols` (#1318).
+
 # tidyr 1.2.0
 
 ## Pivoting